refactor: MealInfoView
